refactor: moved run_with_* test utils in a test_utils mod
