+++ crypto/src/proposal.rs
@@ -84,7 +84,7 @@ pub mod proposal_tests {
     use openmls::prelude::*;
     use wasm_bindgen_test::wasm_bindgen_test;
 
-    use crate::{credential::CredentialSupplier, test_fixture_utils::*, test_utils::run_test_with_central, *};
+    use crate::{credential::CredentialSupplier, test_utils::*, *};
 
     use super::*;
